pyproject.toml
[project]
name = "data-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "deep-translator>=1.11.4",
    "fastapi>=0.100.0",
    "google-generativeai>=0.3.0",
    "gunicorn>=21.2.0",
    "numpy>=1.24.0",
    "onnxruntime>=1.15.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "tensorflow-cpu>=2.13.0",
    "uvicorn[standard]>=0.23.0",
]
